requirements.txt
fastapi
uvicorn
onnxruntime
pandas
shap
xgboost==1.7.6
numpy<2.0.0
google-cloud-storage